+++ src/builder.rs
@@ -1,6 +1,5 @@
 use crate::actor::{ActorTrait, Handler};
 use crate::actor_config::{ActorConfig, RestartPolicy};
-use crate::actor_ref::ActorHandler;
 use crate::config::prelude::DEFAULT_POOL;
 use crate::prelude::ActorRef;
 use crate::system::ActorSystem;