+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - move Actor Messages into separate directory
 - move Mailbox to separate file
+- move ActorHandler to separate file
 
 # WIP 0.0.8